reorganize and generalize mixins
