--- static/js/src/connected-elements/mixins/mapValueToElement.js
@@ -1,21 +1,0 @@
-export default {
-  props: {
-    name: {
-      type: String,
-      required: true,
-    },
-
-    storeModule: {
-      type: String,
-      required: true,
-    },
-  },
-
-  computed: {
-    value() {
-      const getter =
-        this.$store.getters[`${this.storeModule}/valueByKey`];
-      return getter(this.name);
-    },
-  },
-};
