+++ static/js/src/elements/Email.vue
@@ -4,22 +4,18 @@
     :name="name"
     :placeholder="placeholder"
     :class="getCssClasses()"
-    type="text"
+    type="email"
     @input="updateSingleValue($event.target.value)"
   >
 </template>
 
 <script>
-import mapValueToElement from './mixins/mapValueToElement';
-import updateSingleValue from './mixins/updateSingleValue';
+import connectedElement from './mixins/connectedElement';
 
 export default {
-  name: 'TextInput',
+  name: 'Email',
 
-  mixins: [
-    mapValueToElement,
-    updateSingleValue,
-  ],
+  mixins: [connectedElement],
 
   props: {
     placeholder: {
