+++ static/js/src/elements/Level.vue
