+++ static/js/src/elements/mixins/updateStoreValue.js
@@ -1,0 +1,10 @@
+export default {
+  methods: {
+    updateStoreValue({ storeModule, key, value }) {
+      this.$store.dispatch(
+        `${storeModule}/updateValue`,
+        { key, value },
+      );
+    },
+  },
+};
